function_ingest/requirements.txt
requests==2.32.3
google-cloud-storage==2.18.0
flask==3.0.3
functions-framework==3.*